+++ cmd/terramate/cli/cli.go
@@ -222,6 +222,19 @@ type cliSpec struct {
 				} `cmd:"" help:"show drifts"`
 			} `cmd:"" help:"manage cloud drifts"`
 		} `cmd:"" help:"Terramate Cloud commands"`
+
+		Script struct {
+			List struct{} `cmd:"" help:"Show a list of all scripts in the current directory"`
+			Tree struct{} `cmd:"" help:"Show a tree of all scripts in the current directory"`
+			Info struct {
+				Labels []string `arg:"" name:"labels" passthrough:"" help:"Name of the script"`
+			} `cmd:"" help:"Show detailed information about a script"`
+			Run struct {
+				NoRecursive bool     `default:"false" help:"Do not recurse into child stacks"`
+				DryRun      bool     `default:"false" help:"Plan the execution but do not execute it"`
+				Labels      []string `arg:"" name:"labels" passthrough:"" help:"Script to execute"`
+			} `cmd:"" help:"Run script in stacks"`
+		} `cmd:"" help:"Terramate Script commands"`
 	} `cmd:"" help:"Experimental features (may change or be removed in the future)"`
 }
 
@@ -585,6 +598,17 @@ func (c *cli) run() {
 		c.cloudInfo()
 	case "experimental cloud drift show":
 		c.cloudDriftShow()
+	case "experimental script list":
+		c.printScriptList()
+	case "experimental script tree":
+		c.printScriptTree()
+	case "experimental script info <labels>":
+		c.printScriptInfo()
+	case "experimental script run":
+		log.Fatal().Msg("no script specified")
+	case "experimental script run <labels>":
+		c.setupGit()
+		c.runScript()
 	default:
 		log.Fatal().Msg("unexpected command sequence")
 	}